+++ internal/entities/statement.go
@@ -7,9 +7,9 @@ import (
 )
 
 type balance struct {
-	Total         int64     `bson:"total"          json:"total"`
+	Total         int       `bson:"total"          json:"total"`
 	StatementDate time.Time `bson:"statement_date" json:"data_extrato"`
-	Limit         uint64    `bson:"limit"          json:"limite"`
+	Limit         uint      `bson:"limit"          json:"limite"`
 }
 
 type Statement struct {
